requirements.txt
python-telegram-bot[job-queue]>=21.0
aiofiles>=23.1.0
python-dotenv>=1.0.0
requests>=2.31.0
